moved attitude to its own module
